% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gander-package.R
\docType{package}
\name{gander-package}
\alias{gander}
\alias{gander-package}
\title{gander: quick and dirty visualization of your data.frame}
\description{
gander provides a simple and easy to use function that plots all variables in your data frame.
}
\author{
\strong{Maintainer}: Renan Valieris \email{a@b}

}
\keyword{internal}
