% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gander.R
\name{gander}
\alias{gander}
\title{plot all columns of a data.frame}
\usage{
gander(df, ...)
}
\arguments{
\item{df}{A tibble or data.frame}

\item{...}{an optional target column}
}
\description{
plot all columns of a data.frame
}
\examples{
gander(iris)
}
