% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{gander_addin}
\alias{gander_addin}
\title{Run the gander addin}
\usage{
gander_addin()
}
\value{
The underlying ellmer Chat, invisibly. Primarily called for its side effects,
modifying the current RStudio editor based on user input. Will error if
no text is entered in the dialog.
}
\description{
The gander addin is intended to be called using the RStudio addin rather
than explicitly by the user. See \code{vignette("gander", package = "gander")}
to learn more about using the gander addin.
}
\examples{
\dontrun{
# Requires an interactive session, access to the RStudio API,
# and an active connection to an LLM API.
gander_addin()
}

}
