% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{gander_peek}
\alias{gander_peek}
\title{Interface with the previous gander}
\usage{
gander_peek()
}
\value{
The ellmer \code{Chat} object from the last assistant interaction,
or \code{NULL} if no previous interaction exists.
}
\description{
\code{gander_peek()} returns the ellmer \code{Chat} object from the most recent call
to the gander assistant so that you can see what happened under-the-hood.

Note that gander initializes a new chat every time you invoke the addin,
so the token count and conversation history only describes the most recent
interaction with the package.
}
