% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stash.R
\name{last_gander}
\alias{last_gander}
\alias{gander_peek}
\alias{gander_undo}
\alias{gander_retry}
\title{Interface with the previous gander}
\usage{
gander_peek()

gander_undo()

gander_retry()
}
\description{
These functions interface with the most recent call to the
gander assistant:
\itemize{
\item \code{gander_peek()} returns the ellmer \code{Chat} object (so you can see what
happened under-the-hood).
\item \code{gander_undo()} reverts the gander assistant's edits.
\item \code{gander_retry()} submits the same input to the model, replacing the
previous gander with new input.
}
}
