% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickly-package.R
\docType{package}
\name{quickly-package}
\alias{quickly}
\alias{quickly-package}
\title{quickly: LLMs at the tooltip}
\description{
The package brings elmer chats into your RStudio and Positron sessions, editing files directly. After triggering the package addin, type your request and the assistant will stream its response directly into your documents, optionally incorporating context from the active file and its neighboring files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/quickly}
  \item \url{https://simonpcouch.github.io/quickly/}
  \item Report bugs at \url{https://github.com/simonpcouch/quickly/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
