% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{quickly_addin}
\alias{quickly_addin}
\title{Run the quickly addin}
\usage{
quickly_addin()
}
\value{
Nothing. Called for its side effects, modifying the current RStudio editor
based on user input. Will error if no text is entered in the dialog.
}
\description{
Run the quickly addin
}
