% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{quickly_options}
\alias{quickly_options}
\title{Options used by the quickly package}
\description{
The quickly package makes use of two notable user-facing options:
\itemize{
\item \code{.quickly_fn} and \code{.quickly_args} determine the underlying LLM powering the
assistant.
}
}
