% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gander-package.R
\docType{package}
\name{gander-package}
\alias{gander}
\alias{gander-package}
\title{gander: High performance, low friction LLM chat for data scientists}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The package introduces a high performance, low friction LLM chat experience for data scientists. ellmer chats are integrated directly into your RStudio and Positron sessions, automatically incorporating relevant context such as surrounding lines of code, variable summaries (like data frame columns and types), function documentation, and images. After triggering the package addin, type your request and the assistant will stream its response directly into your documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/gander}
  \item \url{https://simonpcouch.github.io/gander/}
  \item Report bugs at \url{https://github.com/simonpcouch/gander/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

}
\keyword{internal}
