% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gander-package.R
\docType{package}
\name{gander-package}
\alias{gander}
\alias{gander-package}
\title{gander: High Performance, Low Friction Large Language Model Chat}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Introduces a 'Copilot'-like completion experience, but it knows how to talk to the objects in your R environment. 'ellmer' chats are integrated directly into your 'RStudio' and 'Positron' sessions, automatically incorporating relevant context from surrounding lines of code and your global environment (like data frame columns and types). Open the package dialog box with a keyboard shortcut, type your request, and the assistant will stream its response directly into your documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/gander}
  \item \url{https://simonpcouch.github.io/gander/}
  \item Report bugs at \url{https://github.com/simonpcouch/gander/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
