% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{gander_options}
\alias{gander_options}
\title{Options used by the gander package}
\description{
The gander package makes use of two notable user-facing options:
\itemize{
\item \code{.gander_fn} and \code{.gander_args} determine the underlying LLM powering the
assistant.
}
}
