% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{gander_options}
\alias{gander_options}
\alias{.gander_fn}
\alias{.gander_args}
\alias{.gander_chat}
\alias{.gander_style}
\title{Options used by the gander package}
\description{
The gander package makes use of a few notable user-facing options.
}
\section{Choosing models}{


gander uses the \code{.gander_chat} option to configure which model powers the
addin. \code{.gander_chat} is a function that returns an ellmer Chat object.
For example, to use OpenAI's GPT-4o-mini, you might write

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(.gander_chat = function() ellmer::chat_claude())
}\if{html}{\out{</div>}}

Paste that code in your \code{.Rprofile} via \code{usethis::edit_r_profile()} to
always use the same model every time you start an R session.

The gander package used to use options \code{.gander_fn} and \code{.gander_args},
but those are deprecated in favor of \code{.gander_chat}.
}

\section{Style/taste}{


By default, gander responses use the following style
conventions: "Use tidyverse style and, when relevant, tidyverse packages. For example, when asked to plot something, use ggplot2, or when asked to transform data, using dplyr and/or tidyr unless explicitly instructed otherwise. " Set the \code{.gander_style} option to
some other string to tailor responses to your own taste, e.g.:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(.gander_style = "Use base R.")
}\if{html}{\out{</div>}}

Paste that code in your
\code{.Rprofile} via \code{usethis::edit_r_profile()} to always use the same style (or
even always begin with some base set of knowledge about frameworks you
work with often) every time you start an R session.
}

