% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{gander_addin}
\alias{gander_addin}
\title{Run the gander addin}
\usage{
gander_addin()
}
\value{
Nothing. Called for its side effects, modifying the current RStudio editor
based on user input. Will error if no text is entered in the dialog.
}
\description{
Run the gander addin
}
